// types/sample.ts
export type MushroomAnalysis = {
  sample_id: string | null;
  sample_name: string | null;
  analytical_method: string | null;
  analytical_lab: string | null;
  date_analyzed: string | null;
  psilocybin_content: string | null;
  psilocin_content: string | null;
  baeocystin_content: string | null;
  nor_baeocystin_content: string | null;
  total_psilocybin_equivalents: string | null;
};